feat(a11y): update accessibility plugin structure and functionality according to new api
- Modified package.json to adjust main and module entry points for better compatibility.
- Added jsxImportSource configuration in TypeScript files to utilize the new plugin-ui components.
- Updated Vite configuration to integrate the new plugin-ui and enhance build processes.
- Refactored the accessibility plugin to streamline the toolbar action rendering and improve user interaction.
- Introduced a new action button component for accessibility audits, enhancing the user experience.
- Removed deprecated code and styles to simplify the plugin's implementation.
