+++ plugins/a11y/package.json
@@ -16,8 +16,8 @@
   "publishConfig": {
     "access": "public"
   },
-  "main": "dist/index.umd.cjs",
-  "module": "dist/index.js",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.es.js",
   "types": "dist/index.d.ts",
   "files": ["dist"],
   "license": "AGPL-3.0-only",
